dock_driver/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
python-jose==3.3.0
python-multipart==0.0.9
pyyaml==6.0.1
pymodbus==3.5.4
paho-mqtt==2.2.1
aiomqtt==1.2.1
aiohttp==3.9.3
asyncio==3.4.3
redis==5.0.1
prometheus-client==0.19.0
